feat: enhance teacher registration by adding name validation and trimming
